pyproject.toml
[tool.poetry]
name = "portalanalytics"
version = "0.1.0"
description = ""
authors = ["Nicholas Lee <nicholas.lee@sagebase.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
toolbox = {git = "https://github.com/nlee-sage/toolbox.git", rev = "main"}
seaborn = "^0.13.0"
matplotlib = "^3.8.1"
plotly = "^5.18.0"
nbformat = "^5.9.2"
dash = "^2.14.1"
tabulate = "^0.9.0"
python-dotenv = "^1.0.0"
cryptography = "^41.0.5"
snowflake-uuid = "^0.1.1"
snowflake-connector-python = {extras = ["pandas", "secure-local-storage"], version = "^3.5.0"}
pyarrow = "^14.0.1"
tqdm = "^4.66.1"
thefuzz = "^0.20.0"
scikit-learn = "^1.3.2"
tensorflow = ">=2.15.0"
bs4 = "^0.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
